requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Azure AI Services
openai==1.6.1
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-core==1.35.0
azure-identity==1.24.0
azure-storage-blob==12.26.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-speech==1.34.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-language-luis==0.7.0
azure-ai-translation-text==1.0.0b1
azure-ai-translation-document==1.0.0
azure-ai-contentsafety==1.0.0

# HTTP and API tools
requests==2.32.5
httpx==0.25.2
aiofiles==23.2.1

# Data validation
pydantic==2.5.3
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# Authentication and security
msal==1.33.0
msal-extensions==1.3.1
PyJWT==2.10.1
cryptography==45.0.7

# Other dependencies
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
idna==3.10
isodate==0.7.2
pycparser==2.22
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
